Serialize partition tuple in manifest files
Add support for serializing the partition tuple in manifest files. This
is a large change with two parts:
* Add the `value` module and the `Value` struct, which represent the
actual value of a field in an Iceberg table. Each `Value` can be
serialized using serde. The serialized value is used in the partition
tuple.
* Dynamically serialize the ManifestEntry and DataFile structs, since
their schema depends on the partitioning spec of the table.